Error en vuelos en la bantcaj Formas de pago en 0 errores de paquetes

+++ EcuaBilling/src/billing/Factura.java
@@ -656,7 +656,7 @@ public class Factura {
 				flgManager.getTrxData(transactionID);
 
 			}
-			if (product.toBill()) {
+			if (product.toBill() && product.getValPag()>0) {
 				if (iD != null) {
 					PreparedStatement _stmt = database.getConnection()
 							.prepareStatement("DELETE FROM AdvFPag WHERE NumDoc = ?");
